refactor app nodes stores into a module
