+++ backend/lores-app-node/src/lib.rs
@@ -1,15 +1,12 @@
 mod backoff;
 mod consumer;
-mod grpc;
-mod local;
 mod node;
-mod outbox;
 mod projection;
-mod store;
+mod stores;
 mod subscription;
 mod types;
 
 pub use node::{AppNode, NodeError};
 pub use projection::ProjectionDb;
-pub use store::StoreError;
+pub use stores::StoreError;
 pub use types::{AppNodeOperation, LoResNodeId, LoResOperationId};
