+++ backend/lores-app-node/src/stores/mod.rs
@@ -81,3 +81,6 @@ pub(crate) trait OperationStore: Send + Sync + 'static {
         })
     }
 }
+pub(crate) mod grpc;
+pub(crate) mod local;
+pub(crate) mod outbox;
